requirements.txt
anyio==4.12.1
asgiref==3.11.1
build==1.4.0
CacheControl==0.14.4
certifi==2026.1.4
charset-normalizer==3.4.4
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
Django==6.0.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djangorestframework-stubs==3.16.8
django-filter==25.1
drf-spectacular==0.29.0
drf-yasg==1.21.15
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.24.3
findpython==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
isort==8.0.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
keyring==25.7.0
more-itertools==10.8.0
msgpack==1.1.2
packaging==26.0
pbs-installer==2026.2.11
pillow==12.1.1
pkginfo==1.12.1.2
platformdirs==4.9.2
poetry==2.3.2
poetry-core==2.3.1
psycopg2-binary==2.9.11
pyproject_hooks==1.2.0
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
ruff==0.15.2
shellingham==1.5.4
sqlparse==0.5.5
tomlkit==0.14.0
trove-classifiers==2026.1.14.14
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
virtualenv==20.38.0
zstandard==0.25.0
stripe==11.8.1
